Create Q1
Write a program that reads three x, y coordinates from the console (STDIN). The first two
coordinates will be used to compute an equation of a straight line passing through each
coordinate. The program should then calculate the shortest distance from the third coordinate
to the line and print the result to the console (STDOUT).